refactor(misc): various updates
* refactor/reorganize/rename modules
* remove common_regression and other utilities no longer needed (functionality matched by flopy get-modflow and modflow6 testing updates)
* update setup.cfg with extra dependency groups and supported python/OS versions
* add pytest.ini to project root
* use pytest --import-mode=importlib in pyproject.toml (as recommended in link above)
* add test CI job
* stub tests
