+++ .github/workflows/ci.yml
@@ -1,0 +1,112 @@
+name: Continuous integration
+on:
+  push:
+    branches:
+      - main
+      - develop
+  pull_request:
+    branches:
+      - main
+      - develop
+jobs:
+  lint:
+    name: Lint
+    runs-on: ubuntu-latest
+    steps:
+
+      - name: Checkout repo
+        uses: actions/checkout@v3
+
+      - name: Setup Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: 3.7
+          cache: 'pip'
+          cache-dependency-path: setup.cfg
+
+      - name: Install Python packages
+        run: |
+          pip3 install .
+          pip3 install ".[lint]"
+
+      - name: Run isort
+        run: isort --verbose --check --diff modflow_devtools
+
+      - name: Run black
+        run: black --check --diff modflow_devtools
+
+      - name: Run flake8
+        run: flake8 --count --show-source modflow_devtools
+
+      - name: Run pylint
+        run: pylint --jobs=0 --errors-only modflow_devtools
+
+  test:
+    name: Test
+    needs: lint
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ ubuntu-latest, macos-latest, windows-latest ]
+        python: [ 3.7, 3.8, 3.9, "3.10" ]
+        exclude:
+          # avoid shutil.copytree infinite recursion bug
+          # https://github.com/python/cpython/pull/17098
+          - python: '3.8.0'
+    defaults:
+      run:
+        shell:
+          bash -l {0}
+    steps:
+
+      - name: Checkout repo
+        uses: actions/checkout@v3
+
+      - name: Setup Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python }}
+          cache: 'pip'
+          cache-dependency-path: setup.cfg
+
+      - name: Install Python packages
+        run: |
+          pip3 install .
+          pip3 install ".[test]"
+
+      - name: Run tests
+        run: pytest -v -n auto --durations 0
+
+  publish:
+    name: Publish
+    needs: test
+    runs-on: ubuntu-latest
+    steps:
+
+      - name: Checkout repo
+        uses: actions/checkout@v3
+
+      - name: Setup Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: 3.7
+
+      - name: Upgrade pip and install build and twine
+        run: |
+          python -m pip install --upgrade pip
+          pip3 install build twine
+
+      - name: Base modflow_devtools installation
+        run: |
+          pip3 --verbose install .
+
+      - name: Print modflow_devtools version
+        run: |
+          python -c "import modflow_devtools; print(modflow_devtools.__version__)"
+
+      - name: Build modflow_devtools, check dist outputs
+        run: |
+          python -m build
+          twine check --strict dist/*
+
