--- .github/workflows/modflow-devtools-linting-install.yml
@@ -1,94 +1,0 @@
-name: modflow_devtools linting/installation
-
-on:
-  schedule:
-    - cron: '0 3 * * 3' # run at 3 AM UTC every Wednesday
-  push:
-    branches:
-      - main
-      - develop
-  pull_request:
-    branches:
-      - develop
-jobs:
-
-
-  modflow-devtools_lint:
-    name: modflow_devtools linting
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout repo
-        uses: actions/checkout@v2.3.4
-
-      - name: Set up Python
-        uses: actions/setup-python@v2
-        with:
-          python-version: 3.9
-
-      - name: Install packages
-        run: |
-          pip install numpy flopy pylint flake8 black requests
-
-      - name: Run isort
-        run: |
-          echo "if isort check fails update isort using"
-          echo "  pip install isort --upgrade"
-          echo "and run"
-          echo "  isort ./modflow_devtools"
-          echo "and then commit the changes."
-          isort --verbose --check --diff ./modflow_devtools
-
-      - name: Run black
-        run: |
-          echo "if black check fails update black using"
-          echo "  pip install black --upgrade"
-          echo "and run"
-          echo "  black ./modflow_devtools"
-          echo "and then commit the changes."
-          black --check --diff ./modflow_devtools
-
-      - name: Run flake8
-        run: flake8 --count --show-source ./modflow_devtools
-
-      - name: Run pylint
-        run: pylint --jobs=0 --errors-only ./modflow_devtools
-
-  modflow-devtools_setup:
-    name: standard installation
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-    defaults:
-      run:
-        shell: bash
-
-    steps:
-
-      # check out repo
-      - name: Checkout repo
-        uses: actions/checkout@v2.3.4
-
-      - name: Setup Python
-        uses: actions/setup-python@v2.2.2
-        with:
-          python-version: 3.9
-
-      - name: Upgrade pip and install build and twine
-        run: |
-          python -m pip install --upgrade pip
-          pip install build twine
-
-      - name: Base modflow_devtools installation
-        run: |
-          pip --verbose install .
-
-      - name: Print modflow_devtools version
-        run: |
-          python -c "import modflow_devtools; print(modflow_devtools.__version__)"
-
-      - name: Build modflow_devtools, check dist outputs
-        run: |
-          python -m build
-          twine check --strict dist/*
-
