+++ modflow_devtools/config.py
@@ -4,7 +4,4 @@ __version__ = "0.0.1"
 __maintainer__ = "Joseph D. Hughes"
 __email__ = "jdhughes@usgs.gov"
 __status__ = "Production"
-__description__ = """
-This is the modflow_devtools package that provides functionality used in the
-MODFLOW development environment.
-"""
+__description__ = """MODFLOW 6 developer tools."""
