+++ modflow_devtools/test/test_build.py
@@ -1,0 +1,2 @@
+def test_meson_build():
+    pass
