+++ modflow_devtools/test/test_misc.py
@@ -1,0 +1,2 @@
+def test_set_dir():
+    pass
