+++ modflow_devtools/test/test_setup.py
@@ -1,0 +1,2 @@
+def test_model_setup():
+    pass
