--- modflow_devtools/utilities/__init__.py
@@ -1,1 +1,0 @@
-# __init__.py
