+++ pytest.ini
@@ -1,0 +1,8 @@
+[pytest]
+addopts = -ra
+python_files =
+    test_*.py
+    *_test*.py
+markers =
+    slow: tests that don't complete in a few seconds
+    meta: run by other tests (e.g. testing fixtures)
