+++ setup.cfg
@@ -16,7 +16,16 @@ classifiers =
     Development Status :: 5 - Production/Stable
     Intended Audience :: Science/Research
     License :: CC0 1.0 Universal (CC0 1.0) Public Domain Dedication
+    Operating System :: Microsoft :: Windows
+    Operating System :: POSIX
+    Operating System :: Unix
+    Operating System :: MacOS
     Programming Language :: Python
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
     Programming Language :: Python :: 3 :: Only
     Topic :: Scientific/Engineering :: Hydrology
 url = https://github.com/MODFLOW-USGS/modflow-devtools
@@ -24,7 +33,7 @@ download_url = https://pypi.org/project/modflow-devtools
 project_urls =
     Documentation = https://mfpymake.readthedocs.io
     Bug Tracker = https://github.com/MODFLOW-USGS/modflow-devtools/issues
-    Source Code = https://github.com/MODFLOW-USGS/modflow-devtools/pymake
+    Source Code = https://github.com/MODFLOW-USGS/modflow-devtools
 
 [options]
 include_package_data = True  # includes files listed in MANIFEST.in
@@ -35,6 +44,21 @@ install_requires =
     numpy
     requests
     flopy
+
+[options.extras_require]
+lint =
+    black
+    cffconvert
+    flake8
+    isort
+    pylint
+test =
+    %(lint)s
+    coverage
+    flaky
+    pytest
+    pytest-cov
+    pytest-xdist
 
 [options.package_data]
 modflow_devtools = utilities/usgsprograms.txt
